Refactor server.py monolith into modular tools/ package
Break the 2189-line server.py into focused modules:
- app.py: FastMCP instance and entry point
- client.py: Cached MonarchMoney client (no keyring hit per call)
- helpers.py: Shared format_transaction, json_error, json_success
- tools/: 10 modules (auth, accounts, transactions, summaries, splits,
  tags, rules, categories, budgets, financial)
- server.py: Backward-compat shim re-exporting all public names

All 39 tools converted from sync+run_async to native async def.
Errors now return consistent JSON via json_error(). bulk_categorize
uses asyncio.gather for concurrency. split_transaction accepts
List[Dict] instead of JSON string. gql imports moved to module-level
in rules.py. Removed dead MonarchConfig class, unused imports, and
asyncio stdlib dep from pyproject.toml. Fixed secure_session.py
cleanup paths to use os.path.expanduser("~").
