+++ pyproject.toml
@@ -26,7 +26,6 @@ dependencies = [
     "keyring>=24.0.0",
     "python-dotenv>=1.0.0",
     "pydantic>=2.0.0",
-    "asyncio>=3.4.3",
     "playwright>=1.40.0",
 ]
