+++ src/monarch_mcp_server/tools/rules.py
@@ -1,0 +1,420 @@
+"""Transaction rules tools with GraphQL queries."""
+
+import logging
+from typing import Any, Dict, List, Optional
+
+from gql import gql
+
+from monarch_mcp_server.app import mcp
+from monarch_mcp_server.client import get_monarch_client
+from monarch_mcp_server.helpers import json_success, json_error
+
+logger = logging.getLogger(__name__)
+
+# ---------------------------------------------------------------------------
+# GraphQL constants
+# ---------------------------------------------------------------------------
+
+GET_TRANSACTION_RULES_QUERY = gql("""
+query GetTransactionRules {
+  transactionRules {
+    id
+    order
+    merchantCriteriaUseOriginalStatement
+    merchantCriteria {
+      operator
+      value
+      __typename
+    }
+    originalStatementCriteria {
+      operator
+      value
+      __typename
+    }
+    merchantNameCriteria {
+      operator
+      value
+      __typename
+    }
+    amountCriteria {
+      operator
+      isExpense
+      value
+      valueRange {
+        lower
+        upper
+        __typename
+      }
+      __typename
+    }
+    categoryIds
+    accountIds
+    categories {
+      id
+      name
+      icon
+      __typename
+    }
+    accounts {
+      id
+      displayName
+      __typename
+    }
+    setMerchantAction {
+      id
+      name
+      __typename
+    }
+    setCategoryAction {
+      id
+      name
+      icon
+      __typename
+    }
+    addTagsAction {
+      id
+      name
+      color
+      __typename
+    }
+    linkGoalAction {
+      id
+      name
+      __typename
+    }
+    setHideFromReportsAction
+    reviewStatusAction
+    recentApplicationCount
+    lastAppliedAt
+    __typename
+  }
+}
+""")
+
+CREATE_TRANSACTION_RULE_MUTATION = gql("""
+mutation Common_CreateTransactionRuleMutationV2($input: CreateTransactionRuleInput!) {
+  createTransactionRuleV2(input: $input) {
+    errors {
+      fieldErrors {
+        field
+        messages
+        __typename
+      }
+      message
+      code
+      __typename
+    }
+    __typename
+  }
+}
+""")
+
+UPDATE_TRANSACTION_RULE_MUTATION = gql("""
+mutation Common_UpdateTransactionRuleMutationV2($input: UpdateTransactionRuleInput!) {
+  updateTransactionRuleV2(input: $input) {
+    errors {
+      fieldErrors {
+        field
+        messages
+        __typename
+      }
+      message
+      code
+      __typename
+    }
+    __typename
+  }
+}
+""")
+
+DELETE_TRANSACTION_RULE_MUTATION = gql("""
+mutation Common_DeleteTransactionRule($id: ID!) {
+  deleteTransactionRule(id: $id) {
+    deleted
+    errors {
+      fieldErrors {
+        field
+        messages
+        __typename
+      }
+      message
+      code
+      __typename
+    }
+    __typename
+  }
+}
+""")
+
+# ---------------------------------------------------------------------------
+# Tools
+# ---------------------------------------------------------------------------
+
+
+@mcp.tool()
+async def get_transaction_rules() -> str:
+    """
+    Get all transaction auto-categorization rules from Monarch Money.
+
+    Returns a list of rules with their conditions and actions.
+    Rules automatically categorize transactions based on merchant, amount, etc.
+    """
+    try:
+        client = await get_monarch_client()
+        result = await client.gql_call(
+            operation="GetTransactionRules",
+            graphql_query=GET_TRANSACTION_RULES_QUERY,
+            variables={},
+        )
+
+        rules_list = []
+        for rule in result.get("transactionRules", []):
+            rule_info = {
+                "id": rule.get("id"),
+                "order": rule.get("order"),
+                "merchant_criteria": rule.get("merchantCriteria"),
+                "merchant_name_criteria": rule.get("merchantNameCriteria"),
+                "original_statement_criteria": rule.get("originalStatementCriteria"),
+                "amount_criteria": rule.get("amountCriteria"),
+                "category_ids": rule.get("categoryIds"),
+                "account_ids": rule.get("accountIds"),
+                "use_original_statement": rule.get("merchantCriteriaUseOriginalStatement"),
+                "set_category_action": {
+                    "id": rule.get("setCategoryAction", {}).get("id"),
+                    "name": rule.get("setCategoryAction", {}).get("name"),
+                } if rule.get("setCategoryAction") else None,
+                "set_merchant_action": {
+                    "id": rule.get("setMerchantAction", {}).get("id"),
+                    "name": rule.get("setMerchantAction", {}).get("name"),
+                } if rule.get("setMerchantAction") else None,
+                "add_tags_action": [
+                    {"id": tag.get("id"), "name": tag.get("name")}
+                    for tag in rule.get("addTagsAction", [])
+                ] if rule.get("addTagsAction") else None,
+                "link_goal_action": rule.get("linkGoalAction"),
+                "hide_from_reports_action": rule.get("setHideFromReportsAction"),
+                "review_status_action": rule.get("reviewStatusAction"),
+                "recent_application_count": rule.get("recentApplicationCount"),
+                "last_applied_at": rule.get("lastAppliedAt"),
+            }
+            rules_list.append(rule_info)
+
+        return json_success(rules_list)
+    except Exception as e:
+        return json_error("get_transaction_rules", e)
+
+
+@mcp.tool()
+async def create_transaction_rule(
+    merchant_criteria_operator: Optional[str] = None,
+    merchant_criteria_value: Optional[str] = None,
+    amount_operator: Optional[str] = None,
+    amount_value: Optional[float] = None,
+    amount_is_expense: bool = True,
+    set_category_id: Optional[str] = None,
+    set_merchant_name: Optional[str] = None,
+    add_tag_ids: Optional[List[str]] = None,
+    hide_from_reports: Optional[bool] = None,
+    review_status: Optional[str] = None,
+    account_ids: Optional[List[str]] = None,
+    apply_to_existing: bool = False,
+) -> str:
+    """
+    Create a new transaction auto-categorization rule.
+
+    Rules automatically categorize future transactions based on conditions.
+
+    Args:
+        merchant_criteria_operator: How to match merchant ("eq", "contains")
+        merchant_criteria_value: Merchant name/pattern to match
+        amount_operator: Amount comparison ("gt", "lt", "eq", "between")
+        amount_value: Amount threshold value
+        amount_is_expense: Whether amount is expense (negative) or income
+        set_category_id: Category ID to assign (use get_categories for IDs)
+        set_merchant_name: Merchant name to set on matching transactions
+        add_tag_ids: List of tag IDs to add (use get_tags for IDs)
+        hide_from_reports: Whether to hide matching transactions from reports
+        review_status: Review status to set ("needs_review" or null)
+        account_ids: Limit rule to specific account IDs
+        apply_to_existing: Whether to apply rule to existing transactions
+
+    Returns:
+        Result of rule creation.
+
+    Example:
+        Create rule: "Amazon purchases → Shopping category"
+        create_transaction_rule(
+            merchant_criteria_operator="contains",
+            merchant_criteria_value="amazon",
+            set_category_id="cat_123"
+        )
+    """
+    try:
+        client = await get_monarch_client()
+
+        rule_input: Dict[str, Any] = {
+            "applyToExistingTransactions": apply_to_existing,
+        }
+
+        if merchant_criteria_operator and merchant_criteria_value:
+            rule_input["merchantNameCriteria"] = [{
+                "operator": merchant_criteria_operator,
+                "value": merchant_criteria_value,
+            }]
+
+        if amount_operator and amount_value is not None:
+            rule_input["amountCriteria"] = {
+                "operator": amount_operator,
+                "isExpense": amount_is_expense,
+                "value": amount_value,
+                "valueRange": None,
+            }
+
+        if account_ids:
+            rule_input["accountIds"] = account_ids
+
+        if set_category_id:
+            rule_input["setCategoryAction"] = set_category_id
+        if set_merchant_name:
+            rule_input["setMerchantAction"] = set_merchant_name
+        if add_tag_ids:
+            rule_input["addTagsAction"] = add_tag_ids
+        if hide_from_reports is not None:
+            rule_input["setHideFromReportsAction"] = hide_from_reports
+        if review_status:
+            rule_input["reviewStatusAction"] = review_status
+
+        result = await client.gql_call(
+            operation="Common_CreateTransactionRuleMutationV2",
+            graphql_query=CREATE_TRANSACTION_RULE_MUTATION,
+            variables={"input": rule_input},
+        )
+
+        errors = result.get("createTransactionRuleV2", {}).get("errors")
+        if errors:
+            return json_success({"success": False, "errors": errors})
+
+        return json_success({"success": True, "message": "Rule created successfully"})
+    except Exception as e:
+        return json_error("create_transaction_rule", e)
+
+
+@mcp.tool()
+async def update_transaction_rule(
+    rule_id: str,
+    merchant_criteria_operator: Optional[str] = None,
+    merchant_criteria_value: Optional[str] = None,
+    amount_operator: Optional[str] = None,
+    amount_value: Optional[float] = None,
+    amount_is_expense: bool = True,
+    set_category_id: Optional[str] = None,
+    set_merchant_name: Optional[str] = None,
+    add_tag_ids: Optional[List[str]] = None,
+    hide_from_reports: Optional[bool] = None,
+    review_status: Optional[str] = None,
+    account_ids: Optional[List[str]] = None,
+    apply_to_existing: bool = False,
+) -> str:
+    """
+    Update an existing transaction rule.
+
+    Args:
+        rule_id: The ID of the rule to update (use get_transaction_rules to find IDs)
+        merchant_criteria_operator: How to match merchant ("eq", "contains")
+        merchant_criteria_value: Merchant name/pattern to match
+        amount_operator: Amount comparison ("gt", "lt", "eq", "between")
+        amount_value: Amount threshold value
+        amount_is_expense: Whether amount is expense (negative) or income
+        set_category_id: Category ID to assign
+        set_merchant_name: Merchant name to set
+        add_tag_ids: List of tag IDs to add
+        hide_from_reports: Whether to hide from reports
+        review_status: Review status to set
+        account_ids: Limit rule to specific accounts
+        apply_to_existing: Apply changes to existing transactions
+
+    Returns:
+        Result of rule update.
+    """
+    try:
+        client = await get_monarch_client()
+
+        rule_input: Dict[str, Any] = {
+            "id": rule_id,
+            "applyToExistingTransactions": apply_to_existing,
+        }
+
+        if merchant_criteria_operator and merchant_criteria_value:
+            rule_input["merchantNameCriteria"] = [{
+                "operator": merchant_criteria_operator,
+                "value": merchant_criteria_value,
+            }]
+
+        if amount_operator and amount_value is not None:
+            rule_input["amountCriteria"] = {
+                "operator": amount_operator,
+                "isExpense": amount_is_expense,
+                "value": amount_value,
+                "valueRange": None,
+            }
+
+        if account_ids:
+            rule_input["accountIds"] = account_ids
+
+        if set_category_id:
+            rule_input["setCategoryAction"] = set_category_id
+        if set_merchant_name:
+            rule_input["setMerchantAction"] = set_merchant_name
+        if add_tag_ids:
+            rule_input["addTagsAction"] = add_tag_ids
+        if hide_from_reports is not None:
+            rule_input["setHideFromReportsAction"] = hide_from_reports
+        if review_status:
+            rule_input["reviewStatusAction"] = review_status
+
+        result = await client.gql_call(
+            operation="Common_UpdateTransactionRuleMutationV2",
+            graphql_query=UPDATE_TRANSACTION_RULE_MUTATION,
+            variables={"input": rule_input},
+        )
+
+        errors = result.get("updateTransactionRuleV2", {}).get("errors")
+        if errors:
+            return json_success({"success": False, "errors": errors})
+
+        return json_success({"success": True, "message": "Rule updated successfully"})
+    except Exception as e:
+        return json_error("update_transaction_rule", e)
+
+
+@mcp.tool()
+async def delete_transaction_rule(rule_id: str) -> str:
+    """
+    Delete a transaction rule.
+
+    Args:
+        rule_id: The ID of the rule to delete (use get_transaction_rules to find IDs)
+
+    Returns:
+        Confirmation of deletion.
+    """
+    try:
+        client = await get_monarch_client()
+
+        result = await client.gql_call(
+            operation="Common_DeleteTransactionRule",
+            graphql_query=DELETE_TRANSACTION_RULE_MUTATION,
+            variables={"id": rule_id},
+        )
+
+        delete_result = result.get("deleteTransactionRule", {})
+        if delete_result.get("deleted"):
+            return json_success({"success": True, "message": "Rule deleted successfully"})
+
+        errors = delete_result.get("errors")
+        if errors:
+            return json_success({"success": False, "errors": errors})
+
+        return json_success({"success": False, "message": "Unknown error"})
+    except Exception as e:
+        return json_error("delete_transaction_rule", e)
