+++ tests/test_transactions.py
@@ -4,13 +4,7 @@ import json
 import pytest
 from unittest.mock import AsyncMock, patch, MagicMock
 
-# Mock the monarchmoney module before importing server
-import sys
-sys.modules['monarchmoney'] = MagicMock()
-sys.modules['monarchmoney'].MonarchMoney = MagicMock
-sys.modules['monarchmoney'].RequireMFAException = Exception
-
-from monarch_mcp_server.server import (
+from monarch_mcp_server.tools.transactions import (
     get_transactions_needing_review,
     set_transaction_category,
     update_transaction_notes,
@@ -26,8 +20,8 @@ from monarch_mcp_server.server import (
 class TestGetTransactionsNeedingReview:
     """Tests for get_transactions_needing_review tool."""
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_get_transactions_needs_review_filter(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_get_transactions_needs_review_filter(self, mock_get_client):
         """Test filtering by needs_review flag."""
         mock_client = AsyncMock()
         mock_client.get_transactions.return_value = {
@@ -60,15 +54,15 @@ class TestGetTransactionsNeedingReview:
         }
         mock_get_client.return_value = mock_client
 
-        result = get_transactions_needing_review(needs_review=True)
+        result = await get_transactions_needing_review(needs_review=True)
 
         transactions = json.loads(result)
         assert len(transactions) == 1
         assert transactions[0]["id"] == "txn_1"
         assert transactions[0]["needs_review"] is True
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_get_transactions_uncategorized_filter(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_get_transactions_uncategorized_filter(self, mock_get_client):
         """Test filtering for uncategorized transactions."""
         mock_client = AsyncMock()
         mock_client.get_transactions.return_value = {
@@ -101,7 +95,7 @@ class TestGetTransactionsNeedingReview:
         }
         mock_get_client.return_value = mock_client
 
-        result = get_transactions_needing_review(
+        result = await get_transactions_needing_review(
             needs_review=True,
             uncategorized_only=True
         )
@@ -111,8 +105,8 @@ class TestGetTransactionsNeedingReview:
         assert transactions[0]["id"] == "txn_1"
         assert transactions[0]["category"] is None
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_get_transactions_with_days_filter(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_get_transactions_with_days_filter(self, mock_get_client):
         """Test filtering by days parameter."""
         mock_client = AsyncMock()
         mock_client.get_transactions.return_value = {
@@ -120,15 +114,15 @@ class TestGetTransactionsNeedingReview:
         }
         mock_get_client.return_value = mock_client
 
-        result = get_transactions_needing_review(days=7, needs_review=False)
+        result = await get_transactions_needing_review(days=7, needs_review=False)
 
         # Verify the API was called with date filters
         call_kwargs = mock_client.get_transactions.call_args.kwargs
         assert "start_date" in call_kwargs
         assert "end_date" in call_kwargs
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_get_transactions_full_details(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_get_transactions_full_details(self, mock_get_client):
         """Test that full transaction details are returned."""
         mock_client = AsyncMock()
         mock_client.get_transactions.return_value = {
@@ -153,7 +147,7 @@ class TestGetTransactionsNeedingReview:
         }
         mock_get_client.return_value = mock_client
 
-        result = get_transactions_needing_review(needs_review=True)
+        result = await get_transactions_needing_review(needs_review=True)
 
         transactions = json.loads(result)
         assert len(transactions) == 1
@@ -171,18 +165,19 @@ class TestGetTransactionsNeedingReview:
         assert len(txn["tags"]) == 1
         assert txn["tags"][0]["name"] == "Online"
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_get_transactions_error(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_get_transactions_error(self, mock_get_client):
         """Test error handling."""
         mock_get_client.side_effect = RuntimeError("Auth needed")
 
-        result = get_transactions_needing_review()
+        result = await get_transactions_needing_review()
 
-        assert "Error getting transactions" in result
-        assert "Auth needed" in result
+        data = json.loads(result)
+        assert data["error"] is True
+        assert "Auth needed" in data["message"]
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_get_transactions_empty(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_get_transactions_empty(self, mock_get_client):
         """Test when no transactions match criteria."""
         mock_client = AsyncMock()
         mock_client.get_transactions.return_value = {
@@ -190,7 +185,7 @@ class TestGetTransactionsNeedingReview:
         }
         mock_get_client.return_value = mock_client
 
-        result = get_transactions_needing_review()
+        result = await get_transactions_needing_review()
 
         transactions = json.loads(result)
         assert len(transactions) == 0
@@ -199,8 +194,8 @@ class TestGetTransactionsNeedingReview:
 class TestSetTransactionCategory:
     """Tests for set_transaction_category tool."""
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_set_category_with_mark_reviewed(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_set_category_with_mark_reviewed(self, mock_get_client):
         """Test setting category and marking as reviewed."""
         mock_client = AsyncMock()
         mock_client.update_transaction.return_value = {
@@ -214,7 +209,7 @@ class TestSetTransactionCategory:
         }
         mock_get_client.return_value = mock_client
 
-        result = set_transaction_category(
+        result = await set_transaction_category(
             transaction_id="txn_123",
             category_id="cat_456",
             mark_reviewed=True
@@ -230,14 +225,14 @@ class TestSetTransactionCategory:
         data = json.loads(result)
         assert "updateTransaction" in data
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_set_category_without_mark_reviewed(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_set_category_without_mark_reviewed(self, mock_get_client):
         """Test setting category without marking as reviewed."""
         mock_client = AsyncMock()
         mock_client.update_transaction.return_value = {"updateTransaction": {}}
         mock_get_client.return_value = mock_client
 
-        set_transaction_category(
+        await set_transaction_category(
             transaction_id="txn_123",
             category_id="cat_456",
             mark_reviewed=False
@@ -247,27 +242,29 @@ class TestSetTransactionCategory:
         call_kwargs = mock_client.update_transaction.call_args.kwargs
         assert "needs_review" not in call_kwargs
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_set_category_error(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_set_category_error(self, mock_get_client):
         """Test error handling."""
         mock_get_client.side_effect = RuntimeError("API error")
 
-        result = set_transaction_category("txn_123", "cat_456")
+        result = await set_transaction_category("txn_123", "cat_456")
 
-        assert "Error setting category" in result
+        data = json.loads(result)
+        assert data["error"] is True
+        assert "API error" in data["message"]
 
 
 class TestUpdateTransactionNotes:
     """Tests for update_transaction_notes tool."""
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_update_notes_simple(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_update_notes_simple(self, mock_get_client):
         """Test updating notes without receipt URL."""
         mock_client = AsyncMock()
         mock_client.update_transaction.return_value = {"updateTransaction": {}}
         mock_get_client.return_value = mock_client
 
-        update_transaction_notes(
+        await update_transaction_notes(
             transaction_id="txn_123",
             notes="Business lunch with client"
         )
@@ -275,14 +272,14 @@ class TestUpdateTransactionNotes:
         call_kwargs = mock_client.update_transaction.call_args.kwargs
         assert call_kwargs["notes"] == "Business lunch with client"
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_update_notes_with_receipt(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_update_notes_with_receipt(self, mock_get_client):
         """Test updating notes with receipt URL."""
         mock_client = AsyncMock()
         mock_client.update_transaction.return_value = {"updateTransaction": {}}
         mock_get_client.return_value = mock_client
 
-        update_transaction_notes(
+        await update_transaction_notes(
             transaction_id="txn_123",
             notes="Office supplies",
             receipt_url="https://drive.google.com/file/abc123"
@@ -291,21 +288,23 @@ class TestUpdateTransactionNotes:
         call_kwargs = mock_client.update_transaction.call_args.kwargs
         assert call_kwargs["notes"] == "[Receipt: https://drive.google.com/file/abc123] Office supplies"
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_update_notes_error(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_update_notes_error(self, mock_get_client):
         """Test error handling."""
         mock_get_client.side_effect = RuntimeError("API error")
 
-        result = update_transaction_notes("txn_123", "test")
+        result = await update_transaction_notes("txn_123", "test")
 
-        assert "Error updating notes" in result
+        data = json.loads(result)
+        assert data["error"] is True
+        assert "API error" in data["message"]
 
 
 class TestMarkTransactionReviewed:
     """Tests for mark_transaction_reviewed tool."""
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_mark_reviewed_success(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_mark_reviewed_success(self, mock_get_client):
         """Test marking transaction as reviewed."""
         mock_client = AsyncMock()
         mock_client.update_transaction.return_value = {
@@ -315,7 +314,7 @@ class TestMarkTransactionReviewed:
         }
         mock_get_client.return_value = mock_client
 
-        result = mark_transaction_reviewed(transaction_id="txn_123")
+        result = await mark_transaction_reviewed(transaction_id="txn_123")
 
         call_kwargs = mock_client.update_transaction.call_args.kwargs
         assert call_kwargs["transaction_id"] == "txn_123"
@@ -324,27 +323,29 @@ class TestMarkTransactionReviewed:
         data = json.loads(result)
         assert "updateTransaction" in data
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_mark_reviewed_error(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_mark_reviewed_error(self, mock_get_client):
         """Test error handling."""
         mock_get_client.side_effect = RuntimeError("API error")
 
-        result = mark_transaction_reviewed("txn_123")
+        result = await mark_transaction_reviewed("txn_123")
 
-        assert "Error marking reviewed" in result
+        data = json.loads(result)
+        assert data["error"] is True
+        assert "API error" in data["message"]
 
 
 class TestBulkCategorizeTransactions:
     """Tests for bulk_categorize_transactions tool."""
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_bulk_categorize_all_success(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_bulk_categorize_all_success(self, mock_get_client):
         """Test successful bulk categorization of all transactions."""
         mock_client = AsyncMock()
         mock_client.update_transaction.return_value = {"updateTransaction": {}}
         mock_get_client.return_value = mock_client
 
-        result = bulk_categorize_transactions(
+        result = await bulk_categorize_transactions(
             transaction_ids=["txn_1", "txn_2", "txn_3"],
             category_id="cat_123",
             mark_reviewed=True
@@ -359,8 +360,8 @@ class TestBulkCategorizeTransactions:
         # Verify update_transaction was called 3 times
         assert mock_client.update_transaction.call_count == 3
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_bulk_categorize_partial_failure(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_bulk_categorize_partial_failure(self, mock_get_client):
         """Test bulk categorization with some failures."""
         mock_client = AsyncMock()
         # First call succeeds, second fails, third succeeds
@@ -371,7 +372,7 @@ class TestBulkCategorizeTransactions:
         ]
         mock_get_client.return_value = mock_client
 
-        result = bulk_categorize_transactions(
+        result = await bulk_categorize_transactions(
             transaction_ids=["txn_1", "txn_2", "txn_3"],
             category_id="cat_123"
         )
@@ -383,14 +384,14 @@ class TestBulkCategorizeTransactions:
         assert len(data["errors"]) == 1
         assert data["errors"][0]["transaction_id"] == "txn_2"
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_bulk_categorize_without_mark_reviewed(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_bulk_categorize_without_mark_reviewed(self, mock_get_client):
         """Test bulk categorization without marking as reviewed."""
         mock_client = AsyncMock()
         mock_client.update_transaction.return_value = {"updateTransaction": {}}
         mock_get_client.return_value = mock_client
 
-        bulk_categorize_transactions(
+        await bulk_categorize_transactions(
             transaction_ids=["txn_1"],
             category_id="cat_123",
             mark_reviewed=False
@@ -399,13 +400,13 @@ class TestBulkCategorizeTransactions:
         call_kwargs = mock_client.update_transaction.call_args.kwargs
         assert "needs_review" not in call_kwargs
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_bulk_categorize_empty_list(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_bulk_categorize_empty_list(self, mock_get_client):
         """Test bulk categorization with empty transaction list."""
         mock_client = AsyncMock()
         mock_get_client.return_value = mock_client
 
-        result = bulk_categorize_transactions(
+        result = await bulk_categorize_transactions(
             transaction_ids=[],
             category_id="cat_123"
         )
@@ -415,24 +416,26 @@ class TestBulkCategorizeTransactions:
         assert data["successful"] == 0
         assert data["failed"] == 0
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_bulk_categorize_client_error(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_bulk_categorize_client_error(self, mock_get_client):
         """Test error when client cannot be obtained."""
         mock_get_client.side_effect = RuntimeError("Auth needed")
 
-        result = bulk_categorize_transactions(
+        result = await bulk_categorize_transactions(
             transaction_ids=["txn_1"],
             category_id="cat_123"
         )
 
-        assert "Error in bulk categorization" in result
+        data = json.loads(result)
+        assert data["error"] is True
+        assert "Auth needed" in data["message"]
 
 
 class TestSearchTransactions:
     """Tests for search_transactions tool."""
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_search_with_text(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_search_with_text(self, mock_get_client):
         """Test text search."""
         mock_client = AsyncMock()
         mock_client.get_transactions.return_value = {
@@ -453,7 +456,7 @@ class TestSearchTransactions:
         }
         mock_get_client.return_value = mock_client
 
-        result = search_transactions(search="Amazon")
+        result = await search_transactions(search="Amazon")
 
         call_kwargs = mock_client.get_transactions.call_args.kwargs
         assert call_kwargs["search"] == "Amazon"
@@ -461,8 +464,8 @@ class TestSearchTransactions:
         transactions = json.loads(result)
         assert len(transactions) == 1
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_search_with_multiple_filters(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_search_with_multiple_filters(self, mock_get_client):
         """Test search with multiple filters."""
         mock_client = AsyncMock()
         mock_client.get_transactions.return_value = {
@@ -470,7 +473,7 @@ class TestSearchTransactions:
         }
         mock_get_client.return_value = mock_client
 
-        search_transactions(
+        await search_transactions(
             search="coffee",
             start_date="2024-01-01",
             end_date="2024-01-31",
@@ -489,8 +492,8 @@ class TestSearchTransactions:
         assert call_kwargs["has_notes"] is False
         assert call_kwargs["is_recurring"] is True
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_search_returns_full_details(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_search_returns_full_details(self, mock_get_client):
         """Test that search returns full transaction details."""
         mock_client = AsyncMock()
         mock_client.get_transactions.return_value = {
@@ -518,7 +521,7 @@ class TestSearchTransactions:
         }
         mock_get_client.return_value = mock_client
 
-        result = search_transactions()
+        result = await search_transactions()
 
         transactions = json.loads(result)
         txn = transactions[0]
@@ -527,21 +530,23 @@ class TestSearchTransactions:
         assert txn["is_recurring"] is False
         assert len(txn["tags"]) == 1
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_search_error(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_search_error(self, mock_get_client):
         """Test error handling."""
         mock_get_client.side_effect = RuntimeError("API error")
 
-        result = search_transactions(search="test")
+        result = await search_transactions(search="test")
 
-        assert "Error searching transactions" in result
+        data = json.loads(result)
+        assert data["error"] is True
+        assert "API error" in data["message"]
 
 
 class TestGetTransactionDetails:
     """Tests for get_transaction_details tool."""
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_get_details_success(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_get_details_success(self, mock_get_client):
         """Test successful retrieval of transaction details."""
         mock_client = AsyncMock()
         mock_client.get_transaction_details.return_value = {
@@ -556,7 +561,7 @@ class TestGetTransactionDetails:
         }
         mock_get_client.return_value = mock_client
 
-        result = get_transaction_details(transaction_id="txn_123")
+        result = await get_transaction_details(transaction_id="txn_123")
 
         mock_client.get_transaction_details.assert_called_once_with(
             transaction_id="txn_123"
@@ -565,21 +570,23 @@ class TestGetTransactionDetails:
         data = json.loads(result)
         assert "getTransaction" in data
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_get_details_error(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_get_details_error(self, mock_get_client):
         """Test error handling."""
         mock_get_client.side_effect = RuntimeError("Transaction not found")
 
-        result = get_transaction_details("txn_invalid")
+        result = await get_transaction_details("txn_invalid")
 
-        assert "Error getting transaction details" in result
+        data = json.loads(result)
+        assert data["error"] is True
+        assert "Transaction not found" in data["message"]
 
 
 class TestDeleteTransaction:
     """Tests for delete_transaction tool."""
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_delete_success(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_delete_success(self, mock_get_client):
         """Test successful transaction deletion."""
         mock_client = AsyncMock()
         mock_client.delete_transaction.return_value = {
@@ -587,7 +594,7 @@ class TestDeleteTransaction:
         }
         mock_get_client.return_value = mock_client
 
-        result = delete_transaction(transaction_id="txn_123")
+        result = await delete_transaction(transaction_id="txn_123")
 
         mock_client.delete_transaction.assert_called_once_with(
             transaction_id="txn_123"
@@ -596,21 +603,23 @@ class TestDeleteTransaction:
         data = json.loads(result)
         assert data["deleteTransaction"]["deleted"] is True
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_delete_error(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_delete_error(self, mock_get_client):
         """Test error handling."""
         mock_get_client.side_effect = RuntimeError("Cannot delete")
 
-        result = delete_transaction("txn_123")
+        result = await delete_transaction("txn_123")
 
-        assert "Error deleting transaction" in result
+        data = json.loads(result)
+        assert data["error"] is True
+        assert "Cannot delete" in data["message"]
 
 
 class TestGetRecurringTransactions:
     """Tests for get_recurring_transactions tool."""
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_get_recurring_success(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_get_recurring_success(self, mock_get_client):
         """Test successful retrieval of recurring transactions."""
         mock_client = AsyncMock()
         mock_client.get_recurring_transactions.return_value = {
@@ -634,7 +643,7 @@ class TestGetRecurringTransactions:
         }
         mock_get_client.return_value = mock_client
 
-        result = get_recurring_transactions()
+        result = await get_recurring_transactions()
 
         recurring = json.loads(result)
         assert len(recurring) == 1
@@ -642,8 +651,8 @@ class TestGetRecurringTransactions:
         assert recurring[0]["stream"]["merchant"] == "Netflix"
         assert recurring[0]["stream"]["frequency"] == "monthly"
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_get_recurring_with_dates(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_get_recurring_with_dates(self, mock_get_client):
         """Test with custom date range."""
         mock_client = AsyncMock()
         mock_client.get_recurring_transactions.return_value = {
@@ -651,7 +660,7 @@ class TestGetRecurringTransactions:
         }
         mock_get_client.return_value = mock_client
 
-        get_recurring_transactions(
+        await get_recurring_transactions(
             start_date="2024-02-01",
             end_date="2024-02-29"
         )
@@ -660,11 +669,13 @@ class TestGetRecurringTransactions:
         assert call_kwargs["start_date"] == "2024-02-01"
         assert call_kwargs["end_date"] == "2024-02-29"
 
-    @patch('monarch_mcp_server.server.get_monarch_client')
-    def test_get_recurring_error(self, mock_get_client):
+    @patch('monarch_mcp_server.tools.transactions.get_monarch_client')
+    async def test_get_recurring_error(self, mock_get_client):
         """Test error handling."""
         mock_get_client.side_effect = RuntimeError("API error")
 
-        result = get_recurring_transactions()
+        result = await get_recurring_transactions()
 
-        assert "Error getting recurring transactions" in result
+        data = json.loads(result)
+        assert data["error"] is True
+        assert "API error" in data["message"]
